+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+**v0.1.1 - 09-15-2026**
+
+- Bumped flugo to v0.2.1 (macOS & AppImage fixes)
+- Android build workflow fix
+
+
 **v0.1.0 - 09-12-2026**
 
 - Initial commit